feat: use_trash flag added

+++ internal/wiper.go
@@ -14,6 +14,7 @@ type Wiper struct {
 	WipeOutPattern []string `json:"wipe_out_pattern,omitempty"  mapstructure:"wipe_out_pattern"  yaml:"wipe_out_pattern"`
 	ExcludeDir     []string `json:"exclude_dir,omitempty" mapstructure:"exclude_dir" yaml:"exclude_dir"`
 	BaseDir        string   `json:"base_dir,omitempty" mapstructure:"base_dir" yaml:"base_dir"`
+	UseTrash       bool     `json:"use_trash,omitempty" mapstructure:"use_trash" yaml:"use_trash"`
 }
 
 func GetInstance() *Wiper {
@@ -27,6 +28,21 @@ func (w Wiper) WipeFiles(dir string) error {
 
 	if dir == "" {
 		dir = w.BaseDir
+	}
+
+	home, err := os.UserHomeDir()
+	if err != nil {
+		return err
+	}
+	trash := path.Join(home, ".Trash")
+	if w.UseTrash {
+		if _, err := os.Stat(trash); err != nil {
+			eslog.Debugf("%s not existing creating it.")
+			err := os.Mkdir(trash, 0700)
+			if err != nil {
+				return err
+			}
+		}
 	}
 
 	if entries, err := os.ReadDir(dir); err != nil {
@@ -45,8 +61,17 @@ func (w Wiper) WipeFiles(dir string) error {
 
 					return matcher.Match([]byte(entry.Name()))
 				}) {
-					err := os.Remove(path.Join(dir, entry.Name()))
-					return err
+					if w.UseTrash {
+						err := os.Rename(path.Join(dir, entry.Name()), path.Join(trash, entry.Name()))
+						if err != nil {
+							return err
+						}
+					} else {
+						err := os.Remove(path.Join(dir, entry.Name()))
+						if err != nil {
+							return err
+						}
+					}
 				} else {
 					eslog.Debug("Skipping", entry.Name())
 				}
